[Import Tx] Add SubmitContent for Classic
