+++ src/app/(sidebar)/transaction/styles.scss
@@ -1,4 +1,5 @@
-@use "../../../../styles/utils.scss" as *;
+@use "../../../styles/utils.scss" as *;
+@use "../../../styles/signatures-table.scss";
 
 .BuildTransaction {
   &__layout {
